Must use BlockPos.containing for float-valued coordinates

+++ src/main/java/dev/huskcasaca/effortless/buildmodifier/mirror/Mirror.java
@@ -40,7 +40,7 @@ public class Mirror implements Modifier {
     private static void coordinateMirrorX(MirrorSettings mirrorSettings, BlockPos oldBlockPos, HashSet<BlockPos> coordinates) {
         //find mirror position
         double x = mirrorSettings.position.x + (mirrorSettings.position.x - oldBlockPos.getX() - 0.5);
-        BlockPos newBlockPos = new BlockPos(x, oldBlockPos.getY(), oldBlockPos.getZ());
+        BlockPos newBlockPos = new BlockPos((int)x, oldBlockPos.getY(), oldBlockPos.getZ());
         coordinates.add(newBlockPos);
 
         if (mirrorSettings.mirrorY) coordinateMirrorY(mirrorSettings, newBlockPos, coordinates);
@@ -50,7 +50,7 @@ public class Mirror implements Modifier {
     private static void coordinateMirrorY(MirrorSettings mirrorSettings, BlockPos oldBlockPos, HashSet<BlockPos> coordinates) {
         //find mirror position
         double y = mirrorSettings.position.y + (mirrorSettings.position.y - oldBlockPos.getY() - 0.5);
-        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), y, oldBlockPos.getZ());
+        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), (int)y, oldBlockPos.getZ());
         coordinates.add(newBlockPos);
 
         if (mirrorSettings.mirrorZ) coordinateMirrorZ(mirrorSettings, newBlockPos, coordinates);
@@ -59,7 +59,7 @@ public class Mirror implements Modifier {
     private static void coordinateMirrorZ(MirrorSettings mirrorSettings, BlockPos oldBlockPos, HashSet<BlockPos> coordinates) {
         //find mirror position
         double z = mirrorSettings.position.z + (mirrorSettings.position.z - oldBlockPos.getZ() - 0.5);
-        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), oldBlockPos.getY(), z);
+        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), oldBlockPos.getY(), (int)z);
         coordinates.add(newBlockPos);
     }
 
@@ -91,7 +91,7 @@ public class Mirror implements Modifier {
     private static void blockStateMirrorX(Player player, MirrorSettings mirrorSettings, BlockPos oldBlockPos, BlockState oldBlockState, Container bagInventory, ItemStack itemStack, InteractionHand hand, Map<BlockPos, BlockState> blockStates, List<ItemStack> itemStacks) {
         //find mirror position
         double x = mirrorSettings.position.x + (mirrorSettings.position.x - oldBlockPos.getX() - 0.5);
-        BlockPos newBlockPos = new BlockPos(x, oldBlockPos.getY(), oldBlockPos.getZ());
+        BlockPos newBlockPos = new BlockPos((int)x, oldBlockPos.getY(), oldBlockPos.getZ());
 
         //Randomizer bag synergy
 //		if (bagInventory != null) {
@@ -115,7 +115,7 @@ public class Mirror implements Modifier {
     private static void blockStateMirrorY(Player player, MirrorSettings mirrorSettings, BlockPos oldBlockPos, BlockState oldBlockState, Container bagInventory, ItemStack itemStack, InteractionHand hand, Map<BlockPos, BlockState> blockStates, List<ItemStack> itemStacks) {
         //find mirror position
         double y = mirrorSettings.position.y + (mirrorSettings.position.y - oldBlockPos.getY() - 0.5);
-        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), y, oldBlockPos.getZ());
+        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), (int)y, oldBlockPos.getZ());
 
         //Randomizer bag synergy
 //		if (bagInventory != null) {
@@ -137,7 +137,7 @@ public class Mirror implements Modifier {
     private static void blockStateMirrorZ(Player player, MirrorSettings mirrorSettings, BlockPos oldBlockPos, BlockState oldBlockState, Container bagInventory, ItemStack itemStack, InteractionHand hand, Map<BlockPos, BlockState> blockStates, List<ItemStack> itemStacks) {
         //find mirror position
         double z = mirrorSettings.position.z + (mirrorSettings.position.z - oldBlockPos.getZ() - 0.5);
-        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), oldBlockPos.getY(), z);
+        BlockPos newBlockPos = new BlockPos(oldBlockPos.getX(), oldBlockPos.getY(), (int)z);
 
         //Randomizer bag synergy
 //		if (bagInventory != null) {
